player: Prevent infinite cycle when setting volume
As added in comment, since we use integer truncations of floating point
operations throughout the code to calculate the volume, we are only
accurate to within one unit. The minimum increment used in the
application seems to be 2 for now, so this fix seems sufficient.
Otherwise, you have to use rounding everywhere.

Fixes: #119